refactor(file-exclusion): move file exclusion logic to MyUtil class
Centralize file exclusion logic in MyUtil to avoid code duplication and improve maintainability. Also move refreshStartPage functionality to MyUtil for consistency.
