+++ src/views/startpagesettingtab.ts
@@ -5,6 +5,7 @@ import { t } from "@/i18n";
 import StyleSettingsModal from "@/views/stylesettingsmodal";
 import SearchExclusionModal from "@/views/searchexclusionmodal";
 import PinnedNotesModal from "@/views/pinnednotesmodal";
+import { MyUtil } from "@/utils/myutil";
 
 export class StartPageSettingTab extends PluginSettingTab {
 	plugin: StartPagePlugin;
@@ -14,15 +15,6 @@ export class StartPageSettingTab extends PluginSettingTab {
 	constructor(app: App, plugin: StartPagePlugin) {
 		super(app, plugin);
 		this.plugin = plugin;
-	}
-
-	private refreshStartPage() {
-		const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_START_PAGE);
-		leaves.forEach((leaf) => {
-			if (leaf.view instanceof StartPageView) {
-				leaf.view.renderContent();
-			}
-		});
 	}
 
 	private updateTitleNavigationBar() {
@@ -80,7 +72,7 @@ export class StartPageSettingTab extends PluginSettingTab {
 				toggle.onChange(async (value) => {
 					this.plugin.settings.includeAllFilesInRecent = value;
 					await this.plugin.saveSettings();
-					this.refreshStartPage();
+					MyUtil.refreshStartPage(this.app);
 				});
 			});
 
@@ -95,7 +87,7 @@ export class StartPageSettingTab extends PluginSettingTab {
 				dropdown.onChange(async (value) => {
 					this.plugin.settings.recentNotesLimit = parseInt(value);
 					await this.plugin.saveSettings();
-					this.refreshStartPage();
+					MyUtil.refreshStartPage(this.app);
 				});
 			});
     }
@@ -126,7 +118,7 @@ export class StartPageSettingTab extends PluginSettingTab {
 				toggle.onChange(async (value) => {
 					this.plugin.settings.showStatBar = value;
 					await this.plugin.saveSettings();
-					this.refreshStartPage();
+					MyUtil.refreshStartPage(this.app);
 				});
 			});
     }
@@ -140,7 +132,7 @@ export class StartPageSettingTab extends PluginSettingTab {
                     .setButtonText(t("style_settings_open_button"))
                     .onClick(() => {
                         new StyleSettingsModal(this.app, this.plugin, () => {
-                            this.refreshStartPage();
+                            MyUtil.refreshStartPage(this.app);
                         }).open();
                     });
             });
@@ -180,6 +172,7 @@ export class StartPageSettingTab extends PluginSettingTab {
                     .onClick(() => {
                         new SearchExclusionModal(this.app, this.plugin, () => {
                             this.display();
+							MyUtil.refreshStartPage(this.app);
                         }).open();
                     });
             });
@@ -223,7 +216,7 @@ export class StartPageSettingTab extends PluginSettingTab {
 					}
 					await this.plugin.saveSettings();
 					this.updateFootComponentDisabledState();
-					this.refreshStartPage();
+					MyUtil.refreshStartPage(this.app);
 				});
 			});
 
@@ -237,7 +230,7 @@ export class StartPageSettingTab extends PluginSettingTab {
 					this.plugin.settings.useRandomFooterText = value;
 					await this.plugin.saveSettings();
 					this.updateFootComponentDisabledState();
-					this.refreshStartPage();
+					MyUtil.refreshStartPage(this.app);
 				});
 			});
 
@@ -251,7 +244,7 @@ export class StartPageSettingTab extends PluginSettingTab {
 					this.plugin.settings.customFooterText = value;
 					await this.plugin.saveSettings();
 					this.updateFootComponentDisabledState();
-					this.refreshStartPage();
+					MyUtil.refreshStartPage(this.app);
 				});
 			});
     }
@@ -279,7 +272,7 @@ export class StartPageSettingTab extends PluginSettingTab {
                     .setCta()
                     .onClick(() => {
                         new PinnedNotesModal(this.app, this.plugin, () => {
-                            this.refreshStartPage();
+                            MyUtil.refreshStartPage(this.app);
                             this.display();
                         }).open();
                     });
